regime-trader/requirements.txt
hmmlearn>=0.3.0
alpaca-trade-api>=3.0.0
alpaca-py>=0.8.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
ta>=0.10.0
scikit-learn>=1.3.0
pyyaml>=6.0
python-dotenv>=1.0.0
websocket-client>=1.6.0
schedule>=1.2.0
rich>=13.0.0
